added the brochure for the membership drive.  updated the become a member button to point to the new brochure 'we help win more cases', which then links directly to the applicaiton form and from there they can view the membership elegibility rules. also added the page routes for the 20 dui myths by leanard stamm

+++ src/Saw/Model/Drive.php
@@ -88,9 +88,11 @@ class Drive extends Model {
     	$this->removeByCriteria(array('belongsTo'=>$this->belongsTo));
     	
     	// delete binary files
-    	foreach ($results as $file):
-			self::$app['upload-mongo']->deleteByCriteria(array('belongsTo'=>$file['_id']));
-		endforeach;
+    	if(!empty($results) && is_array($results)){
+	    	foreach ($results as $file):
+				self::$app['upload-mongo']->deleteByCriteria(array('belongsTo'=>$file['_id']));
+			endforeach;
+		}
 
 	}
 }
